DFGIntegerRangeOptimization is wrong for Upsilon (as 'shadow' nodes are not in SSA form) https://bugs.webkit.org/show_bug.cgi?id=218073
Reviewed by Saam Barati.

JSTests:

The only testcase I managed to get for this bug loops forever when not crashing.
So I use a 1s timeout through --watchdog=1000.

* stress/bounds-checking-in-cold-loop.js: Added.
(true.vm.ftlTrue):

Source/JavaScriptCore:

In DFGIntegerRangeOptimization, when visiting an Upsilon node, we call setEquivalence, that calls setRelationship.
But despite its name, this function does not overwrite a pre-existing relationship, it simply replaces it by an over-approximation of the intersection of the old and new relationship (see the filter method).
Since the old relationship is always (by definition) an over-approximation of this intersection, it will often do nothing at all if it cannot find a closer approximation.
This is a problem specifically for Upsilon nodes, because several of them can store to the same "shadow node" corresponding to a given Phi, so they are the only case where there can already be a completely different relationship for the same nodes (coming from a different Upsilon).

The fix is very simple thanks to a suggestion by Phil: we just remove all relationships referring to the shadow node just before executing an Upsilon.
This is correct since the upsilon effectively kills that shadow node, before making it live again with a different value, and we already aggressively prune the relationshipMaps by liveness.

* dfg/DFGIntegerRangeOptimizationPhase.cpp:


Identifier: 230988@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269107 268f45cc-cd09-0410-ab3c-d52691b4dbfc