REGRESSION (r252689): box-shadow with inset and negative spread renders incorrectly https://bugs.webkit.org/show_bug.cgi?id=209930
Patch by Mason Xiao <me@masonx.ca> on 2020-11-01
Reviewed by Simon Fraser.

Source/WebCore:

r252689 introduced a typo when refactoring the calculation of xOffset.
Instead of subtracting the shadowSpread, it was added.
This caused issues when the shadowSpread was greater than approximately half of the enclosing element's width.

Test: fast/box-shadow/negative-inset-box-shadow.html

* rendering/RenderBoxModelObject.cpp:
(WebCore::RenderBoxModelObject::paintBoxShadow): Change sign of shadowSpread in xOffset calculation for inset

LayoutTests:

* fast/box-shadow/negative-inset-box-shadow-expected.html: Added.
* fast/box-shadow/negative-inset-box-shadow.html: Added.

Identifier: 231101@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269242 268f45cc-cd09-0410-ab3c-d52691b4dbfc