WKContentView might get KVO notifications it doesn't expect https://bugs.webkit.org/show_bug.cgi?id=218558 <rdar://problem/71031491>
Reviewed by Tim Horton.

It's possible that code elsewhere in the system adds KVO
observation to the CALayer of a WKContentView, which is
a problem because our implementation of observeValue only
expects calls that it registered for. Fix this by passing
a context to the registration and passing the observation
call onto the superclass if it isn't the expected context.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView setUpInteraction]):
(-[WKContentView observeValueForKeyPath:ofObject:change:context:]):

Identifier: 231216@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269372 268f45cc-cd09-0410-ab3c-d52691b4dbfc