[watchOS] Don't allow managed preference to enable legacy sync XHRs on watchOS https://bugs.webkit.org/show_bug.cgi?id=218548 <rdar://problem/68417252>
Patch by Luming Yin <luming_yin@apple.com> on 2020-11-04
Reviewed by Geoffrey Garen.

The managed preference to allow legacy sync XHRs during page dismissal is not applicable to watchOS, but we
still attempt to dlopen ManagedConfiguration to get the managed configuration value. This redundant check
blocks the main thread and leads to hangs.

We now compile out this code for the watchOS platforms.

* Shared/WebPreferencesDefaultValues.cpp:
(WebKit::defaultDisallowSyncXHRDuringPageDismissalEnabled):
* Shared/WebPreferencesDefaultValues.h:
* Shared/ios/WebPreferencesDefaultValuesIOS.mm:

Identifier: 231218@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269374 268f45cc-cd09-0410-ab3c-d52691b4dbfc