[ews] Speed up ews status bubble loading speed https://bugs.webkit.org/show_bug.cgi?id=214821
Reviewed by Jonathan Bedard.

* CISupport/ews-app/ews/views/statusbubble.py:
(StatusBubble): This significantly reduces the calculations done to estimate the position in queue. The position is queue
is calculated based on patches sent to buildbot in last DAYS_TO_CHECK days, and results received from buildbot. However, in
case any queue has a huge backlog (of more than 1 day), the position in queue might not be accurate. Slight trade-off for
speed vs accuracy.


Identifier: 231346@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269555 268f45cc-cd09-0410-ab3c-d52691b4dbfc