[macOS] Set preference for overridden languages in the WebContent process after entering the sandbox. https://bugs.webkit.org/show_bug.cgi?id=218097 Source/WebKit:
<rdar://problem/70586545>

Reviewed by Alexey Proskuryakov.

Currently, the preference for overridden languages in the WebContent process is set before entering the sandbox,
which leaves behind an open connection to opendirectoryd. This preference should be set after entering the
sandbox to avoid this. This patch sets the preference after the initialization function has been executed.

API test: WebKit.OverrideAppleLanguagesPreference

* Shared/EntryPointUtilities/Cocoa/XPCService/XPCServiceMain.mm:
(WebKit::setAppleLanguagesPreference):
(WebKit::XPCServiceEventHandler):
(WebKit::XPCServiceMain):
* WebProcess/com.apple.WebProcess.sb.in:

Source/WTF:

Reviewed by Alexey Proskuryakov.

Disable CFPrefs direct mode for older versions of macOS. In practice, it was already disabled since a mach connection to cfprefsd
was established before entering the sandbox on older versions on macOS, but that is fixed in this patch.

* wtf/PlatformEnableCocoa.h:

Tools:

Reviewed by Alexey Proskuryakov.

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKit/AccessibilityReduceMotion.mm:
* TestWebKitAPI/Tests/WebKit/EnableAccessibility.mm:
* TestWebKitAPI/Tests/WebKit/OverrideAppleLanguagesPreference.mm: Added.
(TEST):
* TestWebKitAPI/Tests/WebKit/PreferenceChanges.mm:

LayoutTests:

Reviewed by Alexey Proskuryakov.

Add test for sandbox access to cfprefsd.daemon, and mark cfprefsd tests as failing for Mojave and Catalina, since CFPrefs
direct mode is not enabled on those versions of macOS.

* fast/sandbox/mac/sandbox-mach-lookup-expected.txt:
* fast/sandbox/mac/sandbox-mach-lookup.html:
* platform/mac-catalina/fast/sandbox: Added.
* platform/mac-catalina/fast/sandbox/mac: Added.
* platform/mac-catalina/fast/sandbox/mac/sandbox-mach-lookup-expected.txt: Added.
* platform/mac-mojave/fast/sandbox: Added.
* platform/mac-mojave/fast/sandbox/mac: Added.
* platform/mac-mojave/fast/sandbox/mac/sandbox-mach-lookup-expected.txt: Added.


Identifier: 231375@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269584 268f45cc-cd09-0410-ab3c-d52691b4dbfc