[WPE][TestWebKitAPI] Wait for RenderingState::Started to avoid flakiness https://bugs.webkit.org/show_bug.cgi?id=218841
Reviewed by Adrian Perez de Castro.

The external audio rendering test asked JS to play the video but
checked the condition right away. As with other tests, the full side
effects of the JS call might not have propagated, causing the
assertion to fail. This became especially visible after the build with
r269579.

So, like the other conditions in the test, add a method to wait for
the condition.

* TestWebKitAPI/Tests/WebKitGLib/TestWebKitWebView.cpp:
(AudioRenderingWebViewTest::waitUntilStarted): Added


Identifier: 231504@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269725 268f45cc-cd09-0410-ab3c-d52691b4dbfc