[iOS][FCR] Add an internal feature flag to enable the new appearance https://bugs.webkit.org/show_bug.cgi?id=218873 <rdar://problem/71345270>
Reviewed by Tim Horton.

Source/WebCore:

Add a feature flag for the iOS form control refresh, so that the
feature can be tested.

A new user-agent stylesheet is introduced for the refresh so that
the new styles can be opted in to at runtime. Only styles which
differ from the existing stylesheet will be added to the new one.

* DerivedSources-input.xcfilelist:
* DerivedSources.make:
* WebCore.xcodeproj/project.pbxproj:
* css/iOSFormControlRefresh.css: Added.
* style/InspectorCSSOMWrappers.cpp:
(WebCore::Style::InspectorCSSOMWrappers::collectDocumentWrappers):
* style/UserAgentStyle.cpp:
(WebCore::Style::UserAgentStyle::ensureDefaultStyleSheetsForElement):
* style/UserAgentStyle.h:

Source/WTF:

Enable the feature at build time on PLATFORM(IOS_FAMILY) and disable
the feature at runtime.

* Scripts/Preferences/WebPreferencesInternal.yaml:
* wtf/PlatformEnable.h:
* wtf/PlatformEnableCocoa.h:


Identifier: 231549@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269775 268f45cc-cd09-0410-ab3c-d52691b4dbfc