PlatformDisplayLibWPE: use eglGetPlatformDisplay when possible https://bugs.webkit.org/show_bug.cgi?id=218941
Reviewed by Carlos Garcia Campos.

Use the existing wpe_renderer_backend_egl API to detect any desired
EGL platform that should be used to retrieve the global EGLDisplay
object. If such a platform is specified, the eglGetPlatformDisplay
entrypoing from the EGL_EXT_platform_base extension should be used
to retrieve the corresponding EGLDisplay.

If the API or the extension is not available, or if the EGLDisplay is
not retrieved this way, we fall back to eglGetDisplay.

* platform/graphics/libwpe/PlatformDisplayLibWPE.cpp:
(WebCore::PlatformDisplayLibWPE::initialize):


Identifier: 231606@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269841 268f45cc-cd09-0410-ab3c-d52691b4dbfc