[Web Animations] Ensure we don't schedule animation udpates when there are no styles to update https://bugs.webkit.org/show_bug.cgi?id=219071
Reviewed by Simon Fraser.

Source/WebCore:

Test: webanimations/scheduling-of-animation-without-keyframes.html

For keyframe effects that don't interpolate any CSS property, run the same logic that we already run
to determine when to schedule the next animation update in the case where we are running accelerated
animations.

* animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::timeToNextTick const):

LayoutTests:

Add a test that creates animations with various empty keyframes parameters to check that we don't
schedule animation update for them.

* webanimations/scheduling-of-animation-without-keyframes-expected.txt: Added.
* webanimations/scheduling-of-animation-without-keyframes.html: Added.


Identifier: 231712@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269963 268f45cc-cd09-0410-ab3c-d52691b4dbfc