[iOS] beforeunload event does not fire in MobileSafari https://bugs.webkit.org/show_bug.cgi?id=219102 <rdar://problem/70550655>
Reviewed by Geoff Garen.

Source/WebCore:

MobileSafari on iOS does not implement WKUIDelegate's runJavaScriptAlertPanelWithMessage because
it never shows any before unload prompt. When the client does not implement this delegate,
Chrome::canRunBeforeUnloadConfirmPanel() returns false and this was causing
FrameLoader::shouldClose() to return early, before even firing the beforeunload event in each
frame. I updated our logic so that we now fire the beforeunload events no matter what and we
merely do not attempt to show the beforeunload prompt when Chrome::canRunBeforeUnloadConfirmPanel()
return false, similarly to what we do when the document does not have a user gesture.

Note that we already fire the pagehide and unload events on iOS so this is not a significant
change in policy.

* loader/FrameLoader.cpp:
(WebCore::FrameLoader::shouldClose):
(WebCore::FrameLoader::dispatchBeforeUnloadEvent):

Tools:

Add API test coverage.

* TestWebKitAPI/Tests/WebKit/beforeunload.html:
* TestWebKitAPI/Tests/WebKitCocoa/ModalAlerts.mm:
(-[UIDelegateWithoutRunBeforeUnload webViewDidClose:]):
(-[BeforeUnloadMessageHandler userContentController:didReceiveScriptMessage:]):
(TEST):


Identifier: 231728@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269983 268f45cc-cd09-0410-ab3c-d52691b4dbfc