Nullptr crash in RenderObject::parent https://bugs.webkit.org/show_bug.cgi?id=218484 <rdar://problem/70985057>
Reviewed by Ryosuke Niwa.

Source/WebCore:

Let's imagine the following scenario:

BODY
  LI contenteditable=true
    DIV

If the current visible selection is on DIV and we try to execute document.execCommand("InsertOrderedList") then
the current code will first try to fix the orphaned LI before inserting a new list. Fixing the orphaned
LI means that a new list tag must be created between BODY and LI. There is one caveat though, and is that the
InsertNodeBeforeCommand requires that the parent of the new node (in this case the BODY) must be richly editable
(something that is not happening in the example above). That's why we need to ensure that this precondition is met
before trying to fix the orphaned list item.

Test: fast/editing/insert-list-in-orphaned-list-item-crash.html

* editing/InsertListCommand.cpp:
(WebCore::InsertListCommand::fixOrphanedListChild): Early return if the parent is not richly editable.
(WebCore::InsertListCommand::doApplyForSingleParagraph): Fixed a typo.

LayoutTests:

Added new test case.

* fast/editing/insert-list-in-orphaned-list-item-crash-expected.txt: Added.
* fast/editing/insert-list-in-orphaned-list-item-crash.html: Added.

Identifier: 231759@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@270018 268f45cc-cd09-0410-ab3c-d52691b4dbfc