[css-flex] Images as flex items should use the overridingLogicalHeight when defined to compute the logical width https://bugs.webkit.org/show_bug.cgi?id=218975
Reviewed by Manuel Rego Casasnovas.

LayoutTests/imported/w3c:

Replaced 19 FAIL by PASS expectations.

* web-platform-tests/css/css-flexbox/image-as-flexitem-size-001v-expected.txt:
* web-platform-tests/css/css-flexbox/image-as-flexitem-size-002-expected.txt:
* web-platform-tests/css/css-flexbox/image-as-flexitem-size-005v-expected.txt:
* web-platform-tests/css/css-flexbox/image-as-flexitem-size-006-expected.txt:
* web-platform-tests/css/css-flexbox/image-as-flexitem-size-007v-expected.txt:

Source/WebCore:

RenderReplaced should use the overridingLogicalHeight whenever defined instead of the specified logical height to compute the logical
width using an intrinsic aspect ratio. The overriding height is set by flex containers that need to stretch/shrink their items.
The current code was not considering this case and thus, the intrinsic (non-stretched) logical height was used to compute the logical width,
meaning that the stretching set by the flexbox container was ignored.

Note that it isn't enough to check that there is an overriding height, we must also check that the replaced element has an intrinsic size.
Replaced elements with intrinsic ratios but without intrinsic sizes are handled in a separate code path (it's actually undefined behaviour).
That's why it isn't enough to verify that the element has an aspect ratio, because most SVG graphics actually have defined intrinsic ratios
but not intrinsic sizes.

This allows us to pass an additional 19 subtests in 5 flexbox WPT tests.

* rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::computeReplacedLogicalWidth const): Use the overriding logical height if defined in presence of a valid aspect ratio.

Identifier: 231812@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@270073 268f45cc-cd09-0410-ab3c-d52691b4dbfc