[css-flexbox] WebKit doesn't preserve aspect ratio when computing cross size of flexed images in auto-height flex container https://bugs.webkit.org/show_bug.cgi?id=209983 <rdar://problem/61288094>
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

* web-platform-tests/css/css-flexbox/flexitem-stretch-image-expected.txt: Replace FAIL by PASS expectations.

Source/WebCore:

Aspect ratio was not preserved in the cross axis because WebKit was stretching the items (as they're auto sized) without considering
the aspect ratio. Instead of letting flexbox code deal with that we basically don't stretch them and let RenderReplaced compute the
cross size based on the aspect ratio as it does with other elements that are not flex items.

This allows us to pass 3 tests from the WPT test suite.

* rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::needToStretchChildLogicalHeight const): Return false for replaced elements with aspect ratio.

LayoutTests:

* TestExpectations: Unskipped 2 tests that are working fine now
* css3/flexbox/flexitem.html: Updated expectations. Aspect ratio must be preserved.

Identifier: 231990@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@270288 268f45cc-cd09-0410-ab3c-d52691b4dbfc